requirements.txt
beautifulsoup4>=4.10.0
numpy>=1.21.4
requests>=2.25.1
termcolor>=1.1.0
gurobipy>=9.5
prettytable>=2.4.0
PyYAML>=6.0